vulkan: query cached swapchain image validity
This is only possible with the VK_KHR_shared_presentable_image extension
(for whatever reason), so we use that to query the status of the
swapchain at a later point in time to make sure the cached images are
still valid.

If this isn't available, we simply don't do any caching.