vulkan: permit non-coherent memory
In theory we could try using coherent memory instead of noncoherent
memory on platforms that support it, but this doesn't seem to be
necessarily in either case, as it isn't a performance regression on any
driver that supports it.

This should allow iGPUs (which only have noncoherent memory) to work.