(WIP) gpu: make thread-safe
For the most part, this is not too exciting, because we need to lock
access to the command buffer / vk_ctx on almost every call. The major
exceptions here are pl_tex_create / pl_buf_create, which are the only
functions we really care about with this refactor.

This means we can now safely create host-mapped buffers with
pl_buf_create even when called cross-thread.

Fixes #68