opengl: add initial implementation
This is missing support for texel buffers (which are considered optional
anyway by the API), and also for some swapchain-related features (which
I'm not sure in what way we even want to represent in our code).

This is mainly here so we can get support for OpenGL-related GPU backend
features in the context of e.g. shader generation.

We also update the GLFW demo program to support using OpenGL. This is
all still very primitive and rather untested.